Generating base java response parser files with imports
