+++ ds3-autogen-java/src/main/java/com/spectralogic/ds3autogen/java/JavaCodeGenerator.java
@@ -23,19 +23,17 @@ import com.spectralogic.ds3autogen.api.models.apispec.Ds3ApiSpec;
 import com.spectralogic.ds3autogen.api.models.apispec.Ds3Request;
 import com.spectralogic.ds3autogen.api.models.apispec.Ds3Type;
 import com.spectralogic.ds3autogen.api.models.docspec.Ds3DocSpec;
-import com.spectralogic.ds3autogen.api.models.enums.Classification;
 import com.spectralogic.ds3autogen.java.converters.ClientConverter;
 import com.spectralogic.ds3autogen.java.generators.requestmodels.*;
 import com.spectralogic.ds3autogen.java.generators.responsemodels.BaseResponseGenerator;
 import com.spectralogic.ds3autogen.java.generators.responsemodels.BulkResponseGenerator;
 import com.spectralogic.ds3autogen.java.generators.responsemodels.CodesResponseGenerator;
 import com.spectralogic.ds3autogen.java.generators.responsemodels.ResponseModelGenerator;
+import com.spectralogic.ds3autogen.java.generators.responseparser.BaseResponseParserGenerator;
+import com.spectralogic.ds3autogen.java.generators.responseparser.ResponseParserGenerator;
 import com.spectralogic.ds3autogen.java.generators.typemodels.*;
 import com.spectralogic.ds3autogen.java.helpers.JavaHelper;
-import com.spectralogic.ds3autogen.java.models.Client;
-import com.spectralogic.ds3autogen.java.models.Model;
-import com.spectralogic.ds3autogen.java.models.Request;
-import com.spectralogic.ds3autogen.java.models.Response;
+import com.spectralogic.ds3autogen.java.models.*;
 import com.spectralogic.ds3autogen.utils.Ds3TypeClassificationUtil;
 import com.spectralogic.ds3autogen.utils.Helper;
 import freemarker.template.*;
@@ -50,6 +48,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import static com.spectralogic.ds3autogen.java.models.Constants.*;
+import static com.spectralogic.ds3autogen.java.utils.JavaModuleUtil.getCommandPackage;
 import static com.spectralogic.ds3autogen.utils.ConverterUtil.*;
 import static com.spectralogic.ds3autogen.utils.Ds3RequestClassificationUtil.*;
 import static com.spectralogic.ds3autogen.utils.Ds3TypeClassificationUtil.*;
@@ -72,7 +71,6 @@ public class JavaCodeGenerator implements CodeGenerator {
 
     private final Configuration config = new Configuration(Configuration.VERSION_2_3_23);
 
-    private Ds3ApiSpec spec;
     private FileUtils fileUtils;
     private Path destDir;
 
@@ -90,12 +88,16 @@ public class JavaCodeGenerator implements CodeGenerator {
             final FileUtils fileUtils,
             final Path destDir,
             final Ds3DocSpec docSpec) throws IOException {
-        this.spec = spec;
         this.fileUtils = fileUtils;
         this.destDir = destDir;
 
         try {
-            generateCommands(docSpec);
+            final ImmutableList<Ds3Request> requests = spec.getRequests();
+            final ImmutableMap<String, Ds3Type> types = removeUnusedTypes(
+                    spec.getTypes(),
+                    spec.getRequests());
+
+            generateCommands(requests, types, docSpec);
         } catch (final TemplateException e) {
             e.printStackTrace();
         }
@@ -106,10 +108,13 @@ public class JavaCodeGenerator implements CodeGenerator {
      * @throws IOException
      * @throws TemplateException
      */
-    private void generateCommands(final Ds3DocSpec docSpec) throws IOException, TemplateException {
-        generateAllRequests(docSpec);
-        generateAllModels();
-        generateClient(docSpec);
+    private void generateCommands(
+            final ImmutableList<Ds3Request> requests,
+            final ImmutableMap<String, Ds3Type> types,
+            final Ds3DocSpec docSpec) throws IOException, TemplateException {
+        generateAllRequests(requests, docSpec);
+        generateAllModels(types);
+        generateClient(requests, docSpec);
     }
 
     /**
@@ -118,11 +123,7 @@ public class JavaCodeGenerator implements CodeGenerator {
      * @throws IOException
      * @throws TemplateException
      */
-    private void generateAllModels() throws IOException, TemplateException {
-        final ImmutableMap<String, Ds3Type> types = removeUnusedTypes(
-                spec.getTypes(),
-                spec.getRequests());
-
+    private void generateAllModels(final ImmutableMap<String, Ds3Type> types) throws IOException, TemplateException {
         if (isEmpty(types)) {
             LOG.info("There were no models to generate");
             return;
@@ -236,13 +237,14 @@ public class JavaCodeGenerator implements CodeGenerator {
     }
 
     /**
-     * Generates all of the Request and Response handlers described within the Ds3ApiSpec
-     * excluding SpectraInternal requests.
+     * Generates the Request Handler, Response Handler, and Response Parser
+     * for all specified requests
      * @throws IOException
      * @throws TemplateException
      */
-    private void generateAllRequests(final Ds3DocSpec docSpec) throws IOException, TemplateException {
-        final ImmutableList<Ds3Request> requests = spec.getRequests();
+    private void generateAllRequests(
+            final ImmutableList<Ds3Request> requests,
+            final Ds3DocSpec docSpec) throws IOException, TemplateException {
         if (isEmpty(requests)) {
             LOG.info("There were no requests to generate");
             return;
@@ -250,7 +252,60 @@ public class JavaCodeGenerator implements CodeGenerator {
         for (final Ds3Request request : requests) {
             generateRequest(request, docSpec);
             generateResponse(request);
+            generateResponseParser(request);
         }
+    }
+
+    /**
+     * Generates the Response Parser code for the specified Ds3Request
+     */
+    private void generateResponseParser(final Ds3Request ds3Request) throws IOException, TemplateException {
+        final Template tmpl = getResponseParserTemplate(ds3Request);
+
+        final ResponseParser responseParser = toResponseParser(ds3Request);
+        final Path responsePath = toResponseParserPath(responseParser.getName());
+
+        LOG.info("Getting outputstream for file:" + responsePath.toString());
+
+        try (final OutputStream outStream = fileUtils.getOutputFile(responsePath);
+             final Writer writer = new OutputStreamWriter(outStream)) {
+            tmpl.process(responseParser, writer);
+        }
+    }
+
+    /**
+     * Converts a file name into the path containing said file within the client path
+     * @param fileName The name of a file
+     * @return The client path to the given file
+     */
+    protected Path toResponseParserPath(final String fileName) {
+        return destDir.resolve(
+                baseProjectPath.resolve(
+                        Paths.get(RESPONSE_PARSER_PACKAGE_PATH.replace(".", "/") + "/" + fileName + ".java")));
+    }
+
+    /**
+     * Retrieves the response parser template used to generate the specified request
+     */
+    protected Template getResponseParserTemplate(final Ds3Request ds3Request) throws IOException {
+        //TODO special case as necessary
+        return config.getTemplate("responseparser/response_parser_template.ftl");
+    }
+
+    /**
+     * Converts a Ds3Request into a Response Parser model
+     */
+    protected static ResponseParser toResponseParser(final Ds3Request ds3Request) {
+        final ResponseParserGenerator<?> generator = getResponseParserGenerator(ds3Request);
+        return generator.generate(ds3Request, RESPONSE_PARSER_PACKAGE_PATH);
+    }
+
+    /**
+     * Retrieves the response parser generator used to generate the specified request
+     */
+    protected static ResponseParserGenerator<?> getResponseParserGenerator(final Ds3Request ds3Request) {
+        //TODO special case as necessary
+        return new BaseResponseParserGenerator();
     }
 
     /**
@@ -259,8 +314,9 @@ public class JavaCodeGenerator implements CodeGenerator {
      * @throws IOException
      * @throws TemplateException
      */
-    private void generateClient(final Ds3DocSpec docSpec) throws IOException, TemplateException {
-        final ImmutableList<Ds3Request> requests = spec.getRequests();
+    private void generateClient(
+            final ImmutableList<Ds3Request> requests,
+            final Ds3DocSpec docSpec) throws IOException, TemplateException {
         if (isEmpty(requests)) {
             LOG.info("Not generating client: no requests.");
             return;
@@ -321,15 +377,16 @@ public class JavaCodeGenerator implements CodeGenerator {
      * @param ds3Request A Ds3Request
      * @return A Response
      */
-    private Response toResponse(final Ds3Request ds3Request) {
+    private static Response toResponse(final Ds3Request ds3Request) {
         final ResponseModelGenerator<?> modelGenerator = getResponseTemplateModelGenerator(ds3Request);
         return modelGenerator.generate(ds3Request, getCommandPackage(ds3Request));
     }
-    
+
+    //TODO add test once request generation refactor is completed
     /**
      * Retrieves the associated response generator for the specified Ds3Request
      */
-    private ResponseModelGenerator<?> getResponseTemplateModelGenerator(final Ds3Request ds3Request) {
+    protected static ResponseModelGenerator<?> getResponseTemplateModelGenerator(final Ds3Request ds3Request) {
         if (isAllocateJobChunkRequest(ds3Request)
                 || isHeadObjectRequest(ds3Request)
                 || isHeadBucketRequest(ds3Request)) {
@@ -368,27 +425,6 @@ public class JavaCodeGenerator implements CodeGenerator {
             return config.getTemplate("response/get_object_response_template.ftl");
         }
         return config.getTemplate("response/response_template.ftl");
-    }
-
-    /**
-     * Gets the command package suitable for the given Ds3Request. SpectraDs3 commands
-     * have a separate package, as do notifications.
-     * @param ds3Request A Ds3Request
-     * @return The command package that is suitable for the given Ds3Request
-     */
-    private static String getCommandPackage(final Ds3Request ds3Request) {
-        final StringBuilder builder = new StringBuilder();
-        builder.append(COMMANDS_PACKAGE_PATH);
-        if (ds3Request.getClassification() == Classification.spectrads3) {
-            builder.append(SPECTRA_DS3_PACKAGE);
-        }
-        if (ds3Request.getClassification() == Classification.spectrainternal) {
-            builder.append(SPECTRA_INTERNAL_PACKAGE);
-        }
-        if (isNotificationRequest(ds3Request)) {
-            builder.append(NOTIFICATION_PACKAGE);
-        }
-        return builder.toString();
     }
 
     /**
@@ -435,7 +471,7 @@ public class JavaCodeGenerator implements CodeGenerator {
     /**
      * Retrieves the associated request generator for the specified Ds3Request
      */
-    private static RequestModelGenerator<?> getTemplateModelGenerator(final Ds3Request ds3Request) {
+    protected static RequestModelGenerator<?> getTemplateModelGenerator(final Ds3Request ds3Request) {
         if (hasStringRequestPayload(ds3Request)) {
             return new StringRequestPayloadGenerator();
         }
