+++ ds3-autogen-java/src/main/java/com/spectralogic/ds3autogen/java/models/ResponseParser.java
@@ -1,0 +1,60 @@
+/*
+ * ******************************************************************************
+ *   Copyright 2014-2016 Spectra Logic Corporation. All Rights Reserved.
+ *   Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *   this file except in compliance with the License. A copy of the License is located at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file.
+ *   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *   CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *   specific language governing permissions and limitations under the License.
+ * ****************************************************************************
+ */
+
+package com.spectralogic.ds3autogen.java.models;
+
+import com.google.common.collect.ImmutableSet;
+
+public class ResponseParser {
+
+    private final String name;
+    private final String responseName;
+    private final String packageName;
+    private final String parentClass;
+    private final ImmutableSet<String> imports;
+
+    public ResponseParser(
+            final String name,
+            final String responseName,
+            final String packageName,
+            final String parentClass,
+            final ImmutableSet<String> imports) {
+        this.name = name;
+        this.responseName = responseName;
+        this.packageName = packageName;
+        this.parentClass = parentClass;
+        this.imports = imports;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getPackageName() {
+        return packageName;
+    }
+
+    public String getParentClass() {
+        return parentClass;
+    }
+
+    public ImmutableSet<String> getImports() {
+        return imports;
+    }
+
+    public String getResponseName() {
+        return responseName;
+    }
+}
