+++ ds3-autogen-java/src/main/resources/tmpls/responseparser/response_parser_template.ftl
@@ -1,0 +1,9 @@
+<#include "../copyright.ftl"/>
+
+package ${packageName};
+
+<#include "../imports.ftl"/>
+
+public class ${name} extends ${parentClass}<${responseName}> {
+    //TODO implement
+}
