+++ ds3-autogen-java/src/test/java/com/spectralogic/ds3autogen/java/test/helpers/Ds3ResponseCodeFixtureTestHelper.java
@@ -1,3 +1,18 @@
+/*
+ * ******************************************************************************
+ *   Copyright 2014-2016 Spectra Logic Corporation. All Rights Reserved.
+ *   Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *   this file except in compliance with the License. A copy of the License is located at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file.
+ *   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *   CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *   specific language governing permissions and limitations under the License.
+ * ****************************************************************************
+ */
+
 package com.spectralogic.ds3autogen.java.test.helpers;
 
 import com.google.common.collect.ImmutableList;
@@ -5,9 +20,9 @@ import com.spectralogic.ds3autogen.api.models.apispec.Ds3ResponseCode;
 import com.spectralogic.ds3autogen.api.models.apispec.Ds3ResponseType;
 
 /**
- * This class provides utilities for testing Response Converter
+ * This class provides utilities for creating Ds3ResponseCodes for testing
  */
-public final class BaseResponseGeneratorTestHelper {
+public final class Ds3ResponseCodeFixtureTestHelper {
 
     /**
      * Creates a populated Ds3ResponseCode with a non-error response code
