+++ ds3-autogen-java/src/test/java/com/spectralogic/ds3autogen/java/utils/JavaModuleUtil_Test.java
@@ -1,0 +1,40 @@
+/*
+ * ******************************************************************************
+ *   Copyright 2014-2016 Spectra Logic Corporation. All Rights Reserved.
+ *   Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *   this file except in compliance with the License. A copy of the License is located at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file.
+ *   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *   CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *   specific language governing permissions and limitations under the License.
+ * ****************************************************************************
+ */
+
+package com.spectralogic.ds3autogen.java.utils;
+
+import org.junit.Test;
+
+import static com.spectralogic.ds3autogen.java.utils.JavaModuleUtil.getCommandPackage;
+import static com.spectralogic.ds3autogen.testutil.Ds3ModelFixtures.createBucketRequest;
+import static com.spectralogic.ds3autogen.testutil.Ds3ModelFixtures.createBucketSpectraS3Request;
+import static com.spectralogic.ds3autogen.testutil.Ds3ModelFixtures.getRequestCreateNotification;
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+public class JavaModuleUtil_Test {
+
+    @Test
+    public void getCommandPackage_Test() {
+        assertThat(getCommandPackage(createBucketRequest()),
+                is("com.spectralogic.ds3client.commands"));
+
+        assertThat(getCommandPackage(createBucketSpectraS3Request()),
+                is("com.spectralogic.ds3client.commands.spectrads3"));
+
+        assertThat(getCommandPackage(getRequestCreateNotification()),
+                is("com.spectralogic.ds3client.commands.spectrads3.notifications"));
+    }
+}
