+++ ds3-autogen-java/src/test/java/com/spectralogic/ds3autogen/java/utils/TestGeneratedCodeHelper.java
@@ -7,13 +7,16 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
+import static com.spectralogic.ds3autogen.utils.NormalizingContractNamesUtil.toResponseName;
+import static com.spectralogic.ds3autogen.utils.NormalizingContractNamesUtil.toResponseParserName;
 import static org.mockito.Mockito.when;
 
 public final class TestGeneratedCodeHelper {
 
     public final static String CLIENT_PATH = "./ds3-sdk/src/main/java/com/spectralogic/ds3client/";
+    public final static String PARSER_PATH = CLIENT_PATH + "commands/parsers/";
 
-    public enum PathType { REQUEST, RESPONSE }
+    public enum PathType { REQUEST, RESPONSE, RESPONSE_PARSER }
 
     private TestGeneratedCodeHelper() { }
 
@@ -25,7 +28,11 @@ public final class TestGeneratedCodeHelper {
                 builder.append(requestName);
                 break;
             case RESPONSE:
-                builder.append(requestName.replace("Request", "Response"));
+                builder.append(toResponseName(requestName));
+                break;
+            case RESPONSE_PARSER:
+                builder.append(toResponseParserName(requestName));
+                break;
         }
         builder.append(".java");
         return builder.toString();
